Move user database error classes
- Move user database error classes under new 'user' folder for better project structure
- Update references t0 user database error classes
