+++ src/shared/errors/database/user/userExistsError.ts
@@ -4,7 +4,7 @@
  * @format
  */
 
-import DatabaseError from './databaseError';
+import DatabaseError from '../databaseError';
 
 /**
  * Error for when a user already exists in the database.
